# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "listmonk mailing list manager REST API.",
        "title": "listmonk API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/reload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Reload the application",
                "operationId": "reloadApp",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/bounces": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bounces"
                ],
                "summary": "List bounce records",
                "operationId": "listBounces",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by campaign ID",
                        "name": "campaign_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by bounce source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order results by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PageResults"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bounces"
                ],
                "summary": "Delete bounce records",
                "operationId": "deleteBounces",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Delete all bounces",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Bounce IDs to delete",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/bounces/blocklist": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bounces"
                ],
                "summary": "Blocklist all bounced subscribers",
                "operationId": "blocklistBouncedSubscribers",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/bounces/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bounces"
                ],
                "summary": "Get a bounce record",
                "operationId": "getBounce",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bounce ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Bounce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bounces"
                ],
                "summary": "Delete a bounce record",
                "operationId": "deleteBounce",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bounce ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/campaigns/analytics/{type}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Get campaign analytics",
                "operationId": "getCampaignAnalytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analytics type (views, clicks, bounces, links)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Campaign IDs",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CampaignAnalyticsCount"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/campaigns/auto-paused": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Get currently auto-paused campaigns",
                "operationId": "getAutoPausedCampaigns",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_knadh_listmonk_internal_core.AutoPausedCampaign"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CampaignAnalyticsCount": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "PageResults": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "query": {
                    "type": "string"
                },
                "results": {},
                "search": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "github_com_knadh_listmonk_internal_core.AutoPausedCampaign": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pause_reason": {
                    "$ref": "#/definitions/null.String"
                }
            }
        },
...
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Operator API only (/api/operator/*) - static bearer token from [operator].token, e.g. \"Bearer \u003ctoken\u003e\". Fork-only, off by default.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}